% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set for fish}
\format{
## `dtt`
A data frame with 144 rows and 5 variables:
\describe{
  \item{LENGTH}{len of fish}
  \item{WEIGHT}{weight of fish}
  ...
}
}
\source{
text book
}
\usage{
ddt
}
\description{
144 lines of measurements on fish
}
\keyword{datasets}
