% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlikg2.R
\name{maxlikg2}
\alias{maxlikg2}
\title{maxlikelihood2}
\usage{
maxlikg2(theta1, theta2, lfun = "logbinpois", ...)
}
\arguments{
\item{theta1}{first theta to be used}

\item{theta2}{second theta}

\item{lfun}{lfun}

\item{...}{other param}
}
\description{
maxlikelihood2
}
