% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBinomial.R
\name{mybin}
\alias{mybin}
\title{My Binomial}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{sample size}

\item{p}{probability}
}
\description{
My Binomial
}
\examples{
mybin(100, n=20, 0.7)
}
