% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(mu = 10, sigma = 5, b = 6)
}
\arguments{
\item{mu}{= mean of functio}

\item{sigma}{= standard deviation of function}

\item{a}{= input of probability and area to be calculated}
}
\value{
a curve
}
\description{
My Normal Curve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
