% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets for project 1}
\usage{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{num of seats on plane}

\item{gamma}{prob plane will be truly overbooked}

\item{p}{prob of a 'show'}
}
\description{
ntickets for project 1
}
